requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
Django==3.0.6
django-appconf==1.0.4
django-bootstrap3==12.1.0
django-crispy-forms==1.9.0
django-matplotlib==0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-user-accounts==3.0.2
idna==2.9
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.18.4
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
uWSGI==2.0.18
mysqlclient==2.0.1
